@@ -1,31 +1,31 @@
 {
-  "homepage": "http://caleb-distributive.github.io/pretenders-dom6",
+  "homepage": "http://zollqir.github.io/pretenders-dom6",
   "name": "pretenders-dom6",
   "version": "2.00.0",
+  "type": "module",
   "private": true,
   "dependencies": {
-    "sass": "^1.69.0",
-    "ramda": "^0.27.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-scripts": "3.4.1",
-    "react-select": "^5.8.0"
+    "react": "^19.0.0",
+    "react-dom": "^19.0.0",
+    "react-select": "^5.10.0",
+    "sass": "^1.85.0"
+  },
+  "devDependencies": {
+    "@vitejs/plugin-react": "^4.3.4",
+    "gh-pages": "^6.1.1",
+    "vite": "^6.0.0"
   },
   "scripts": {
-    "start": "react-scripts start",
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
     "predeploy": "npm run build",
-    "deploy": "gh-pages -d build",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
+    "deploy": "gh-pages -d dist"
   },
   "browserslist": [
     ">0.2%",
     "not dead",
     "not ie <= 11",
     "not op_mini all"
-  ],
-  "devDependencies": {
-    "gh-pages": "^6.1.1"
-  }
+  ]
 }
